Move keyboard display values to types
