--- src/components/Keyboard/Keyboard.types.ts
@@ -1,6 +1,0 @@
-import { ComponentPropsWithAs } from '../../utils';
-
-export type KeyboardProps<TTag extends React.ElementType> =
-  ComponentPropsWithAs<TTag>;
-
-export const DEFAULT_KEYBOARD_TAG = 'span' as const;
